style(mobile): otomatik lint temizliği — dart fix + withOpacity→withValues
- dart fix --apply (root + example paketleri): 419 fix — prefer_const_constructors
  (~340), unnecessary_non_null_assertion (20), unnecessary_import (16),
  prefer_final_locals, unnecessary_cast, dangling_library_doc_comments vb.
- withOpacity→withValues(alpha:) — 132 kullanım, 43 dosya (Flutter deprecation,
  davranış-eş; iç-içe paren yok, güvenli perl-replace).
- valueColor ölü-field temizliği (staff/team _DetailRow — hiç geçilmiyordu,
  hep null; dart fix param'ı kaldırıp final-init bozuyordu → field+dead
  null-aware kaldırıldı).

analyze: 209→77 issue, 0 error. 627/627 test yeşil (regresyonsuz).

+++ lib/src/core/reporting/reporting_service.dart
@@ -70,12 +70,12 @@ class ReportingService {
         _getPendingNotificationCount(tenantId),
       ]);
 
-      final organizationCount = results[0] as int;
-      final siteCount = results[1] as int;
-      final unitCount = results[2] as int;
-      final activeUserCount = results[3] as int;
-      final activityCount = results[4] as int;
-      final pendingNotificationCount = results[5] as int;
+      final organizationCount = results[0];
+      final siteCount = results[1];
+      final unitCount = results[2];
+      final activeUserCount = results[3];
+      final activityCount = results[4];
+      final pendingNotificationCount = results[5];
 
       // Trend verilerini hesapla
       final previousPeriod = period.getDateRange().previousPeriod;
